new movie/image export defaults

+++ lib/spacetime/gui/windows.py
@@ -302,8 +302,8 @@ class ExportDialog(support.UtilityWindow):
 
 	dpi = traits.Range(low=1, high=10000000, value=72)
 
-	canvas_width = traits.Float(800)
-	canvas_height = traits.Float(600)
+	canvas_width = traits.Float(1024)
+	canvas_height = traits.Float(768)
 	canvas_unit = traits.Enum('px', 'cm', 'inch')
 	figsize = traits.Property(depends_on='canvas_width, canvas_height, canvas_unit')
 
@@ -372,10 +372,10 @@ class MovieDialogMainTab(traits.HasTraits):
 	codec = traits.Str('libx264')
 	ffmpeg_options = traits.Str('-x264opts crf=12 -preset medium -profile main -threads 0')
 
-	frame_width = traits.Int(800)
-	frame_height = traits.Int(600)
+	frame_width = traits.Int(1024)
+	frame_height = traits.Int(768)
 	dpi = traits.Range(low=1, high=10000000, value=72)
-	frame_rate = traits.Int(2)
+	frame_rate = traits.Int(5)
 
 	animation_view = traitsui.View(traitsui.Group(
 		traitsui.Group(
